Fixed includes

+++ src/dataset_creation.hpp
@@ -25,6 +25,6 @@
 
 #include <vector>
 
-std::vector<unsigned long long int> createDataset(unsigned long int tot_num, unsigned int non_zero_percentage);
+std::vector<unsigned long long int> createDataset(unsigned long int tot_num, double non_zero_percentage);
 
 #endif
